src/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.5.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
django-allauth==0.61.1
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-rest-auth==0.9.5
Django==4.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pip==21.3.1
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
requests==2.31.0
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.37.1
